% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilerFactory.R
\name{addProfiler}
\alias{addProfiler}
\title{addProfiler}
\usage{
addProfiler(timed_fun = data.frame(fun=c("read.table"),pkg=c("utils"),prc=c("READ"),typ=c("IO")))
}
\arguments{
\item{timed_fun}{a data.frame containing all functions that should be timed. If empty returns NULL}
}
\value{
override input functions in environment(ExecEnvironment)
}
\description{
This function loops through a data.frame with functions that need to be timed and overrides these function with a timed version of the function in ExecEnvironment.
}

