\name{basic-class}
\alias{pval-class}
\alias{padj-class}
\alias{labels-class}
\alias{stratify-class}
\docType{class}
\title{basic-class of object \code{SimResults}}

\description{
All S4 class for storing results of object of \code{\link{SimResults}}. 
\code{pval} is a matrix or NULL.
\code{padj} is a matrix or NULL.
\code{label} is numeric or NULL.
\code{stratify} is a data.frame or NULL.
}


\seealso{ \code{\link{SimResults}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

