% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchCleaner.R
\name{benchCleaner}
\alias{benchCleaner}
\title{benchCleaner}
\usage{
benchCleaner(target)
}
\arguments{
\item{target}{define what to clean. 
\code{profiles}: clears all profiling records. 
\code{benchmarks}: clears all benchmark records. 
\code{meta}: clears all records about the system. 
\code{warnings}: clears all recorded warnings.}
}
\description{
functions for cleaning data.frame created by benchmarkR
}

