% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchGetter.R
\name{benchGetter}
\alias{benchGetter}
\title{benchGetter}
\usage{
benchGetter(target, indexCol = NULL, returnCol = NULL, selectValue = NULL,
  selectedRunId = NULL, file = NULL)
}
\arguments{
\item{target}{which information to retrieve 
\code{id}: generates a unique ID based on date/time and a random number. 
\code{allprofiles}: returns all records from ExecEnvironment$PROFILES table. 
\code{allbenchmarks}: returns table with all recorded benchmarks. 
\code{profile}: returns 'returnCol' from ExecEnvironment$PROFILES where the value of 'indexCol' is identical to 'selectValue'. 
\code{profilerun}: returns PROFILES table subsetted with \code{runId}. 
\code{warnings}: returns ExecEnvironment$WARNINGS data.frame containing all warnings recorded. 
\code{systemid}: returns the unique systemId that is used to identify this system. 
\code{usedpackages}: returns names of all used/loaded packages within provided input file/script.}

\item{indexCol}{profiling records will be filtered based on content of this column}

\item{returnCol}{content of this column will be returned after filtering}

\item{selectValue}{this value will be compared to \code{indexCol} to filter profiling records}

\item{selectedRunId}{which \code{runId} to use for filtering profiling records}

\item{file}{packages used in this file will be extracted and installed}
}
\description{
Retrieve different saved records or benchmark information
}

