% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkSource.R
\name{benchmarkSource}
\alias{benchmarkSource}
\title{benchmarkSource}
\usage{
benchmarkSource(file, timed_fun = NULL)
}
\arguments{
\item{file}{R script to benchmark}

\item{timed_fun}{a data.frame whith 4 columns. Column 1: function; column 2: package; column 3: process category eg. READ/WRITE but never BENCHMARK; column 4: function type, currently only 'IO'.}
}
\value{
returns a dubble with running time of last benchmark and prints all session benchmark records to console
}
\description{
This script will benchmark the running time of the given input file. 
Time used by functions defined in timed_fun data.frame will be subtracted from total running time.
}
\examples{
library("benchmarkR")
timed_functions <- data.frame(
  timed_fun = c(
    "write.table","write.csv","write.csv2","write.delim2",
    "read.table","read.csv","read.csv2","read.delim","read.delim2",
    "download.file","curlGetHeaders", "fread"
    ),
  package = c(
    "utils","utils","utils","utils","utils","utils",
    "utils","utils","utils", "utils", "base", "data.table"
    ),
  category = c(
    "WRITE","WRITE","WRITE","WRITE","READ","READ",
    "READ","READ","READ", "DOWNLOAD", "DOWNLOAD", "READ"
    ),
  type = c(
    "IO", "IO", "IO", "IO", "IO", "IO",
    "IO", "IO", "IO", "IO", "IO", "IO"
  ),
  stringsAsFactors=FALSE
)

benchmarkSource(file = "./test.R", timed_fun = timed_functions)

benchmarkSource(file = "./test.R")

benchGetter("allprofiles")

benchGetter("allbenchmarks")
}

