\name{BenchMetric-class}
\alias{BenchMetric-class}
\docType{class}
\title{BenchMetric - class}

\description{
A simple S4 class as a basic subclass to build \code{rocX}, \code{powerFDR} and others.
}



\section{Slots}{
  \describe{
    \item{\code{element}:}{A simple list}
  }
}

\seealso{ \code{\link{rocX}}, \code{\link{powerFDR}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

