\name{BenchMetricList-class}
\alias{BenchMetricList-class}
\docType{class}
\title{BenchMetricList - class}

\description{
The object of 'BenchMetricList' class is a container for storing a collection of 'BenchMetric' objects.

}


\seealso{\code{\link{BenchMetric-class}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

