% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchSetter.R
\name{calcComputeTime}
\alias{calcComputeTime}
\title{calcComputeTime}
\usage{
calcComputeTime(runId)
}
\arguments{
\item{runId}{runId of the benchmark/profiling for which running time is to be calculated.}
}
\value{
Running time of the benchmark minus time used by functions provided for profiling.
}
\description{
calculates process running time.
}

