% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchSetter.R
\name{checkSource}
\alias{checkSource}
\title{checkSource}
\usage{
checkSource(file = BenchmarkEnvironment$file,
  runId = BenchmarkEnvironment$runId)
}
\arguments{
\item{file}{name of the file being benchmarked (BenchmarkEnvironment$file)}

\item{runId}{runId of the current benchmark (BenchmarkEnvironment$runId)}
}
\value{
number of direct calls detected (invisble), warning with number of direct calls is printed to console
}
\description{
benchmarkR only read/write commands that are provided to benchmarkSource() are timed. 
Direct calls used by input script (package::function) are not timed and won't be substracted from the total run time. 
This function, therefore, checks if inside input script direct function calls are made using ::/::: notation. 
If this is the case, this function will print warnings for each direct call.
}

