\docType{data}
\name{dat_rocX}
\alias{dat_rocX}

\title{A count-base simulation as an example dataset to test \code{rocX}.}
\description{
  A count-base simulation from Pickrell (HapMap-YRI), 1000 tags, 10 samples (5 control versus 5 treatment) 
}
\usage{data(rocX)
}

\source{
Recount project: http://bowtie-bio.sourceforge.net/recount/ or
http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/
}

\references{
Zhou, Xiaobei and Lindsay, Helen and Robinson, Mark D (2014). Robustly detecting differential expression in RNA sequencing data using observation weights.
\emph{Nucleic Acids Research} doi: 10.1093/nar/gku310.
\url{http://nar.oxfordjournals.org/content/early/2014/04/20/nar.gku310.full}
}


\keyword{datasets}
