\docType{data}
\name{ABRF}
\alias{ABRF}
\alias{pvalAB}
\alias{padjAB}
\alias{labelAB}


\title{DE analysis results of difference between cell lines A and B in Association of Biomolecular Resource Facilities (ABRF) Studies to test \code{rocX}, \code{fdX} and other functions}
\description{
    DE analysis results contain "edgeR", "limma-voom" and "DESeq2". The truth labels are based on Taqman PCR data. The procedure can be found on "script/ABRF.R"
}
\usage{data(ABRF)
}

\source{
http://physiology.med.cornell.edu/faculty/mason/lab/data3/sac2026/ABRF/index.html#
}

\references{
Multi-platform assessment of transcriptome profiling using RNA-seq in the ABRF next-generation sequencing study
\emph{Nature Biotechnology} doi: 10.1038/nbt.2972.
\url{http://www.nature.com/nbt/journal/v32/n9/full/nbt.2972.html}
}


\keyword{datasets}
