\docType{data}
\name{Pickrell}
\alias{Pickrell}

\title{A count-based simulation as an example dataset to test \code{\link{rocX}}, \code{\link{fdX}} and other functions}
\description{
  A count-based simulation from Pickrell (HapMap-YRI), 4000 tags, 10 samples (5 control versus 5 treatment). DE analysis results contain "edgeR", "limma-voom" and "DESeq2". 
}
\usage{data(Pickrell)
}

\source{
Recount project: http://bowtie-bio.sourceforge.net/recount/ or
http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/
}

\references{
Zhou, Xiaobei and Lindsay, Helen and Robinson, Mark D (2014). Robustly detecting differential expression in RNA sequencing data using observation weights.
\emph{Nucleic Acids Research} doi: 10.1093/nar/gku310.
\url{http://nar.oxfordjournals.org/content/early/2014/04/20/nar.gku310.full}
}


\keyword{datasets}
