\docType{data}
\name{PickrellList}
\alias{PickrellList}

\title{A data example set containing 5 sets of p-values and corresponding labels based on the count-based simulation}
\description{
  A data example set containing 5 sets of p-values and corresponding labels based on the count-based simulation. Each count-based simulation from Pickrell (HapMap-YRI), 4000 tags, 10 samples (5 control versus 5 treatment). DE analysis results contain "edgeR", "limma-voom" and "DESeq2". 
}
\usage{data(PickrellList)
}

\source{
Recount project: http://bowtie-bio.sourceforge.net/recount/ or
http://imlspenticton.uzh.ch/robinson_lab/edgeR_robust/
}

\references{
Zhou, Xiaobei and Lindsay, Helen and Robinson, Mark D (2014). Robustly detecting differential expression in RNA sequencing data using observation weights.
\emph{Nucleic Acids Research} doi: 10.1093/nar/gku310.
\url{http://nar.oxfordjournals.org/content/early/2014/04/20/nar.gku310.full}
}

\seealso{ \code{\link{Pickrell}}}

\keyword{datasets}
