% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilerFactory.R
\name{factorsAsStrings}
\alias{factorsAsStrings}
\title{factorsAsStrings}
\usage{
factorsAsStrings(data_frame = data.frame(a=c(1,2),b=c("AA","BB")))
}
\arguments{
\item{data_frame}{input data.frame}
}
\value{
data.frame with all factor columns converted to character columns
}
\description{
converts all Factor columns within input data.frame to Character columns
}

