\name{fdX-class}
\alias{fdX-class}
\docType{class}
\title{fdX - class}

\description{
A simple S4 class for storing results of \code{fdX} for reproducible plots and re-analysis.
}



\section{Slots}{
  \describe{
    \item{\code{element}:}{A list contains: number of selected positive value (number), number of false discoveries (fd), number of selected positive value given a certain threshold (numberX) and number of false discoveries given a certain threshold (fdX).}
  }
}

\seealso{ \code{\link{fdX}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

