% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetup.R
\name{installUsedPackages}
\alias{installUsedPackages}
\title{installUsedPackages}
\usage{
installUsedPackages(file)
}
\arguments{
\item{file}{path to a R script.}
}
\value{
packages loaded within input R script using library() or require() will be installed
}
\description{
installUsedPackages
}
\examples{
installUsedPackages("/tests/test.R")
}

