\name{powerFDR-class}
\alias{powerFDR-class}
\docType{class}
\title{powerFDR - class}

\description{
A simple S4 class for storing results of \code{powerFDR} for reproducible plots and re-analysis.
}



\section{Slots}{
  \describe{
    \item{\code{element}:}{A list contains: false discovery rate (fdr), true positive rate (tpr) and the threshold value (control FDR) assessed...}
  }
}

\seealso{ \code{\link{powerFDR}} }


\author{Charity Law, Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

