\name{precREC-class}
\alias{precREC-class}
\docType{class}
\title{precREC - class}

\description{
A simple S4 class for storing results of \code{precREC} for reproducible plots and re-analysis.
}



\section{Slots}{
  \describe{
    \item{\code{element}:}{A list contains: precision (PREC), true positive rate (tpr) and the threshold value (control precision) assessed...}
  }
}

\seealso{ \code{\link{precREC}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

