% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilerFactory.R
\name{ProfilerFactory}
\alias{ProfilerFactory}
\title{ProfilerFactory}
\usage{
ProfilerFactory("read.table","utils","READ","IO")
}
\arguments{
\item{fun}{function to encapsulate with timing}

\item{pkg}{name of the package that contains this function}

\item{prc}{name for process category that should be used in profile eg. READ or WRITE.}

\item{typ}{type of function eg. IO, DB, or GRAPH. Only IO is currently implemented}
}
\value{
new timed function based on input function
}
\description{
creates a new function by encapsulating input function with code neccessary for timing.
}

