\name{rocXList-class}
\alias{rocXList-class}
\docType{class}
\title{rocXList - class}

\description{
The object of \code{rocXList} class is a container for storing a collection of \code{rocX} objects.

}

\section{Objects from the Class}{Objects can be created by using the \code{rocXList} function.
}


\seealso{\code{\link{rocXList}}, \code{\link{rocX-class}}, \code{\link{rocX}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

