% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchSetter.R
\name{setBenchmark}
\alias{setBenchmark}
\title{setBenchmark}
\usage{
setBenchmark()
}
\value{
Adds the running time of the whole script as a record to ExecEnvironment$BENCHMARKS data.frame
}
\description{
adds last benchmark to benchmark results.
}

