% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchSetter.R
\name{setSystemID}
\alias{setSystemID}
\title{setSystemID}
\usage{
setSystemID()
}
\value{
unique id as character vector and system information are added 
as record to .BenchEnv$META data.frame
}
\description{
Generats a unique ID for the system on which the benchmark
is runned ones on loading of package and stores system information with 
this ID.
}

