
\docType{class}

\name{SimResultsList-class}
\alias{SimResultsList-class}


\title{Class \code{\link{SimResultsList}} }

\description{An S4 Object class \code{\link{SimResultsList}}  to store tore multiple (\code{SimResults}) results.
}


\section{Objects from the class}{Objects can be created by using the \code{\link{SimResultsList}} function.
}





\seealso{ \code{\link{SimResultsList}} }


\author{Xiaobei Zhou and Mark D. Robinson.}


\keyword{SimResultsList}
