\name{basic-class}
\alias{pval-class}
\alias{padj-class}
\alias{labels-class}
\alias{stratify-class}
\docType{class}
\title{basic-class of object \code{SimResults}}

\description{
All S4 class for storing results of object of \code{SimResults}. 
'pval' is a matrix or NULL.
'padj' is a matrix or NULL.
'label' is numeric or NULL.
'label' is a data.frame or NULL.
}


\seealso{ \code{\link{SimResults}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

