% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchDBReport.R
\name{benchDBReport}
\alias{benchDBReport}
\title{benchDBReport}
\usage{
benchDBReport(usr, pwd, host_address, db_name, con_type)
}
\arguments{
\item{usr}{database usrname}

\item{host_address}{IP address of the MySQL server}

\item{db_name}{database name}

\item{con_type}{type of database (currently only mysql supported)}

\item{psw}{database password}
}
\description{
Uploads benchmarking/profiling results tp database
}

