% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchDBReport.R
\name{benchDBReport}
\alias{benchDBReport}
\title{benchDBReport}
\usage{
benchDBReport(usr, pwd, con_str)
}
\arguments{
\item{usr}{database usrname}

\item{con_str}{database connection string jdbc:mysql:location/database/table}

\item{psw}{database password}
}
\description{
Uploads benchmarking/profiling results tp database
}

