% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkSource.R
\name{benchmarkSource}
\alias{benchmarkSource}
\title{Simple Benchmark/Profiling Tool For R Scripts}
\usage{
benchmarkSource(file, timed_fun, runs, loc.src, uses_packrat)
}
\arguments{
\item{file}{R script to benchmark}

\item{timed_fun}{a data frame whith 4 columns. Column 1: function; column 2: 
package; column 3: process category eg. READ/WRITE but never BENCHMARK; 
column 4: function type, currently only 'IO'.}

\item{runs}{(integer) number of pre-runs which are not timed but serve to 
measure warm running time of the last run.}

\item{loc.src}{chracter vector containing location of CRAN like repository
which contains the source code (package_version.tar.gz) of 
used package versions}

\item{uses_packrat}{use if you have packified your script using packrat.
Default is TRUE}
}
\value{
returns a dubble with running time of last benchmark
}
\description{
This script will benchmark the running time of the given input 
file. Time used by functions defined in timed_fun data frame will be 
subtracted from total running time. This benchmarking tools differs to
other benchmarking tools in that it times the running time of longer
R scripts in a defined environment by allowing (version)control over input
data, used packages and scriptfile.
}
\details{
benchmarkSource
}
\author{
Parham Solaimani (Maintainer)

Maarten-Jan Kallen

Alexander Bertram
}
\keyword{benchmark}
\keyword{benchmarking}
\keyword{file}
\keyword{profiling}
\keyword{r}
\keyword{script}
\keyword{timing}

