% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilerFactory.R
\name{factorsAsStrings}
\alias{factorsAsStrings}
\title{factorsAsStrings}
\usage{
factorsAsStrings(df)
}
\arguments{
\item{df}{input object}
}
\value{
Replaces factor columns in list or data frame, otherwise returns input object as is.
}
\description{
converts all Factor columns within input data.frame to Character columns
}

