% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentSetup.R
\name{installUsedPackages}
\alias{installUsedPackages}
\title{installUsedPackages}
\usage{
installUsedPackages(file = "/tests/test.R")
}
\arguments{
\item{file}{path to a R script.}
}
\value{
installs packages loaded within input script.
}
\description{
Installs all packages loaded within a R script using library/require function. This scripts installs all the used packages that are not installed yet
}

