\name{rocXList-class}
\alias{rocXList-class}
\docType{class}
\title{rocX - class}

\description{
The object of 'rocXList' class is a container for storing a collection of
     'rocX' objects.

}


\seealso{ \code{\link{rocX-class}}, \code{\link{rocX}} }


\author{Xiaobei Zhou and Mark D. Robinson}

\keyword{classes}

