% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmark.R
\name{runBenchmark}
\alias{runBenchmark}
\title{runBenchmark}
\usage{
runBenchmark(bench_loc, timed_fun = NULL)
}
\arguments{
\item{bench_loc}{location of directory with subdirectories containing benchmarks}

\item{timed_fun}{data.frame with functions to be profiled}
}
\description{
This function extracts benchmark information, and downloads the necessary data/packages
}

