% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmark.R
\name{runBenchmark}
\alias{runBenchmark}
\title{runBenchmark}
\usage{
runBenchmark(runs, loc.src, dcf, timedFunFile, bench_log, warn_log)
}
\arguments{
\item{runs}{number of runs for warming up. After warmup runs the workflow
will be runned one more time for timing perpose. Default is 0.}

\item{loc.src}{location of CRAN like repository containing the source code
of used package versions. Default is NULL.}

\item{dcf}{information about workflows name and description and 
datafile name, download location, md5hash in Debian Control Format. Default
is BENCHMARK.dcf}

\item{timedFunFile}{file with functions to be profiled. Default is
FUNCTION.PROFILE}

\item{bench_log}{filename to use to store message/warning/errors.
Default is "benchmark.log"}

\item{warn_log}{level of logging to use (choose from -1, "INFO", "WARN", 
"STOP". Default value is -1.}
}
\description{
This function extracts benchmark information, and downloads the necessary data/packages.
}

